+++ src/types/enums/public.ts
@@ -1,0 +1,7 @@
+export enum PUBLIC {
+  ADULTES = "Adultes",
+  ENFANTS = "Enfants/adolescents",
+  BOTH = "Adultes et enfants/adolescents",
+}
+
+export const allPublics = Object.values(PUBLIC);